rename interface -> schemas
