+++ src/sparsify/auto/utils/nm_api.py
@@ -19,7 +19,7 @@ from typing import Tuple
 
 import requests
 
-from sparsify.interface import APIArgs, Metrics, SparsificationTrainingConfig
+from sparsify.schemas import APIArgs, Metrics, SparsificationTrainingConfig
 from sparsify.utils import get_base_url
