+++ tests/sparsify/schemas/test_api.py
@@ -14,7 +14,7 @@
 
 import pytest
 
-from sparsify.interface import APIArgs
+from sparsify.schemas import APIArgs
 
 
 @pytest.mark.parametrize(
